resAb/requirements.txt
aiobotocore==2.23.2
aiodns==3.5.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aioitertools==0.12.0
aiosignal==1.4.0
amqp==5.3.1
annotated-types==0.6.0
anyio==4.10.0
asgiref==3.10.0
asttokens==3.0.1
attrs==25.4.0
billiard==4.2.4
boto3==1.39.8
botocore==1.39.8
Bottleneck==1.4.2
brotlicffi==1.2.0.0
celery==5.6.2
certifi==2026.2.25
cffi==2.0.0
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
comm==0.2.3
contourpy==1.3.3
cycler==0.12.1
debugpy==1.8.16
decorator==5.2.1
distro==1.9.0
Django==5.2.8
django-celery-results==2.5.1
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
executing==2.2.1
fonttools==4.61.1
frozenlist==1.8.0
fsspec==2025.10.0
gunicorn==25.1.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
intel-cmplr-lib-ur==2025.3.2
intel-openmp==2025.3.2
ipykernel==7.1.0
ipython==9.9.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jiter==0.12.0
jmespath==1.0.1
joblib==1.5.3
jupyter_client==8.7.0
jupyter_core==5.9.1
kiwisolver==1.4.9
kombu==5.6.2
matplotlib==3.10.8
matplotlib-inline==0.2.1
mkl==2025.3.1
mkl_fft==2.1.1
mkl_random==1.3.0
mkl-service==2.5.2
multidict==6.7.0
munkres==1.1.4
nest-asyncio==1.6.0
numexpr==2.14.1
numpy==2.4.1
onemkl-license==2025.3.1
openai==2.14.0
packaging==25.0
pandas==2.3.3
parso==0.8.5
pexpect==4.9.0
pgvector==0.4.2
pillow==12.0.0
pip==25.3
platformdirs==4.5.1
ply==3.11
polars==1.39.0
polars-runtime-32==1.39.0
prompt_toolkit==3.0.52
propcache==0.4.1
psutil==7.0.0
psycopg2-binary==2.9.11
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==21.0.0
pycares==4.10.0
pycparser==2.21
pydantic==2.12.4
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.11.0
pyparsing==3.3.1
PyQt6==6.9.1
PyQt6-Qt6==6.9.2
PyQt6_sip==13.10.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
pyzmq==27.1.0
s3fs==2025.10.0
s3transfer==0.13.1
scikit-learn==1.8.0
scipy==1.16.3
setuptools==80.9.0
sip==6.12.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
stack-data==0.6.3
tbb==2022.3.1
tcmlib==1.4.1
threadpoolctl==3.5.0
tomli==2.4.0
tornado==6.5.4
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.15.0
typing-inspection==0.4.2
tzdata==2025.2
tzlocal==5.3.1
umf==1.0.3
unicodedata2==16.0.0
urllib3==2.6.3
uvloop==0.22.1
vine==5.1.0
wcwidth==0.2.14
wheel==0.45.1
whitenoise==6.12.0
wrapt==1.17.0
yarl==1.22.0
